The blocksize is now a long rather than an int.

+++ etomica/data/AccumulatorAverageCovariance.java
@@ -18,7 +18,7 @@ public class AccumulatorAverageCovariance extends AccumulatorAverageFixed {
     public AccumulatorAverageCovariance() {
     }
 
-    public AccumulatorAverageCovariance(int blockSize) {
+    public AccumulatorAverageCovariance(long blockSize) {
         super(blockSize);
     }
